Rename GpsStatusProvider to GnssStatusProvider
We will support more GNSSs than just GPS.
